security/manager/ssl/tests/mochitest/mixedcontent/test_dynDelayedUnsecurePicture.html

<!DOCTYPE HTML>
<html>
<head>
  <title>img.src changes to unsecure test</title>
  <script type="text/javascript" src="/MochiKit/Base.js"></script>
  <script type="text/javascript" src="/MochiKit/DOM.js"></script>
  <script type="text/javascript" src="/MochiKit/Style.js"></script>
  <script type="text/javascript" src="/MochiKit/Signal.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mixedContentTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

  <script class="testbody" type="text/javascript">
  "use strict";

  function runTest()
  {
    isSecurityState("secure");
    window.setTimeout(function() {
      // Don't do this synchronously from onload handler
      document.getElementById("image1").src =
        "http://example.com/tests/security/manager/ssl/tests/mochitest/mixedcontent/moonsurface.jpg";
    }, 0);

    waitForSecurityState("broken", function()
    {
      isSecurityState("broken", "src='http://...' changed to broken");
      finish();
    });
  }

  function afterNavigationTest()
  {
    is(document.getElementById("image1").src,
      "https://example.com/tests/security/manager/ssl/tests/mochitest/mixedcontent/moonsurface.jpg",
      "img.src secure again");
    isSecurityState("secure", "security full after navigation");
    finish();
  }

  </script>
</head>

<body>
  <img id="image1" src="https://example.com/tests/security/manager/ssl/tests/mochitest/mixedcontent/moonsurface.jpg" />
</body>
</html>
